# Reuse session for 60% speedup