--- a/tsconfig.json
+++ b/tsconfig.json
@@ -3,7 +3,7 @@
     "target": "ES2020",
     "useDefineForClassFields": true,
     "module": "ESNext",
-    "lib": ["ES2020", "DOM", "DOM.Iterable","WebWorker"],
+    "lib": ["ES2022", "DOM", "DOM.Iterable","WebWorker"],
     "skipLibCheck": true,
     "allowJs": true,