Hey @k9p5,

vits-web is really awesome!

I've started trying to speed it up a little, currently `predict()` setups a whole new ort session and loads the models each time. If instead you split the process into two, so that you can create a "vits-web session" first, you can get a 60% speedup on the basic example. This is particularly useful for repeated calls, setup the model once, and then repeatedly call it with text chunks for more audio (this is what I want to use it for).

As you can see in the video from my machine it cuts the call time from ~2.7s to ~1.1s, with the ~1.7s during the initiation of the session. Repeat calls then only take ~1s.

I've not added/modified any tests yet as it makes more sense to run this past you first.


https://github.com/diffusion-studio/vits-web/assets/31130/14e5409c-e96a-4e09-9f66-bf5989780b19