@k9p5 2024-07-10
I just noticed that you're also holding the model in memory, was that on purpose?