@k9p5 2024-07-11
Makes sense, it kind of depends on the use case whether to keep the memory footprint as small as possible (my goal) or to minimize runtime. Effectively, the part that is missing from the library are environment variables. I'm gonna copy this strategy from onnix then you can just set something like `tts.env.keepSessionInMemory = true` and maybe this also requires `tts.releaseMemory()`