@samwillis 2024-07-11
Absolutely, and I think vits-web can cover both use cases well. On the minimum runtime side, it's important to be able to start the runtime+model before you need it too, not just keep it around for the next call.

In my refactor the original `predict` function still exists and is just a thin wrapper around the session object, starting it for the single call, then discarding it. The largest part of the memory footprint is the voice model, so I think having a wrapper session object (started with the voiceId) works quite well. It seems to me this API is the most flexible, for example allowing someone to start and use multiple voices at once.

`tts.env.keepSessionInMemory = true` implies that there wouldn't be a way to start the runtime+model pre-emptively, that's very important for the time to first output in some applications. With the thing I'm experimenting with (voice gen in real time of a streaming response from an LLM) I would want to start the model at application start, then when required produce voice output as fast as possible.

Let me know if you are happy with my approach, and if so I can tidy it up and add tests.