@guest271314 2024-07-12
This is what I'm doing to avoid redundancy.

At the top level

```
import { createPiperPhonemize } from "./piper.js";
import * as ort from "./onyx-runtimeweb.js";
```

I'm avoiding fetching dependencies at runtime, and creating Blob URL's, too. 

```
const piperPhonemizeWasm = "./piper_phonemize.wasm"; // URL.createObjectURL(await (await fetch("./piper_phonemize.wasm")).blob());
const piperPhonemizeData = "./piper_phonemize.data"; //URL.createObjectURL(new Blob([await (await fetch("./piper_phonemize.data")).arrayBuffer()]));

async function tts(config) {
  const path = PATH_MAP[config.voiceId];
  const input = JSON.stringify([
    {
      text: config.text.trim(),
    },
  ]);

  ort.env.wasm.numThreads = navigator.hardwareConcurrency;
  ort.env.wasm.wasmPaths = "./";
  const modelConfig =
    (await import(`./${path}.json`, { with: { type: "json" } })).default; //(await fetch(`./${path}.json`)).json();
  const phonemeIds = await new Promise(async (resolve) => {
    const module = await createPiperPhonemize({
      print: (data) => {
        resolve(JSON.parse(data).phoneme_ids);
      },
      printErr: (message) => {
        throw new Error(message);
      },
      locateFile: (url) => {
        if (url.endsWith(".wasm")) return piperPhonemizeWasm;
        if (url.endsWith(".data")) return piperPhonemizeData;
        return url;
      },
    });
    module.callMain([
      "-l",
      modelConfig.espeak.voice,
      "--input",
      input,
      "--espeak_data",
      "/espeak-ng-data",
    ]);
  });
  const sampleRate = modelConfig.audio.sample_rate;
  const noiseScale = modelConfig.inference.noise_scale;
  const lengthScale = modelConfig.inference.length_scale;
  const noiseW = modelConfig.inference.noise_w;
  const modelBlob = URL.createObjectURL(
    new Blob([await (await fetch(`./${path}`)).arrayBuffer()]),
  );
  const session = await ort.InferenceSession.create(modelBlob);
  const feeds = {
    input: new ort.Tensor("int64", phonemeIds, [1, phonemeIds.length]),
    input_lengths: new ort.Tensor("int64", [phonemeIds.length]),
    scales: new ort.Tensor("float32", [noiseScale, lengthScale, noiseW]),
  };
  if (Object.keys(modelConfig.speaker_id_map).length) {
    Object.assign(feeds, {
      sid: new ort.Tensor("int64", [0]),
    });
  }
  const {
    output: { data: pcm },
  } = await session.run(feeds);
  return new Blob([pcm2wav(pcm, 1, sampleRate)], {
    type: "audio/x-wav",
  });
}
```

Next I'm going to use the `Float32Array`'s that are the `pcm` variable directly in an `AudioWorklet`.