@avarayr 2024-09-20
Hey guys, sorry to revive this thread. I'm I right to assume that the recent commit [bdf7f36](https://github.com/diffusionstudio/vits-web/commit/bdf7f3607db04354410e00fd588df560f4b8616e) addresses this issue?