@mikebaldry 2024-11-11
> Hey guys, sorry to revive this thread. I'm I right to assume that the recent commit [bdf7f36](https://github.com/diffusionstudio/vits-web/commit/bdf7f3607db04354410e00fd588df560f4b8616e) addresses this issue?

I think it addresses part of the issue, but it will still create a new session every `predict` and I'm not sure the loading of the imports is the slowest part. That would be creating the `InferenceSession` from the model (and maybe getting the blob, though that's probably cached anyway).

This is very promising for me - I would like to create a session, run many predictions against the session, then close the session when it makes sense for me. 

Maybe it would be good to provide an interface for these operations, then just have the current `predict` do that. It would mean there are high-level and lower-level ways of using it, depending on your use case.