@guest271314 2024-12-20
The fastest is to use `rhasspy/piper` directly.